nddplayground/pyproject.toml
[project]
name = "nddplayground"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "matplotlib>=3.10.8",
    "numpy>=2.4.2",
    "pandas>=3.0.0",
    "pillow>=12.1.1",
    "pyarrow>=23.0.1",
]
